web/php/auth.php now generates a suitable key if one doesn't exist.
Note that any directories specified in the $keyfile path must exist, only the
file itself will be created automatically. The secret/ directory should be used
to store the key (this is the default in web/php/config.php) and the
secret/.gitignore file ensures key data is ignored by git.