Fixed blue top bar showing instead of grey w/ Uni branding.
The web/php/config.php has a $colour var which in a roundabout way is pulled
down by the client side javascript. The live site leaves it unspecified which
produces the grey background & University logo (which we want). Ideally I
suppose play/ should be configured to show a green header as it does with the
current deployment, but this is hardly a large issue.